requirements.txt
Django==1.8.2
django-tinymce==1.5.3
MySQL-python==1.2.5
Pillow==2.8.1
